requirements.txt
# =============================================================================
# ROP Detection Project Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# Deep Learning Framework
tensorflow>=2.10.0

# Image Processing
opencv-python>=4.5.0
Pillow>=9.0.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning Utilities
scikit-learn>=1.0.0

# Data Handling
pandas>=1.3.0

# Configuration
PyYAML>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Explainability
lime>=0.2.0
scikit-image>=0.19.0

# Progress Bar
tqdm>=4.62.0

# Jupyter Support (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: GPU Support (uncomment if needed)
# tensorflow-gpu>=2.10.0

